+++ TagsCloudVisualization_Tests/Program.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using Autofac;
-using TagsCloudVisualization;
-
-namespace TagsCloudVisualization_Tests
+﻿namespace TagsCloudVisualization_Tests
 {
     public static class Program
     {
