+++ TagsCloudVisualization_Tests/TagsCloudVisualization_Tests.csproj
@@ -38,18 +38,29 @@
     <Reference Include="Autofac, Version=4.9.0.0, Culture=neutral, PublicKeyToken=17863af14b0044da">
       <HintPath>..\packages\Autofac.4.9.0-beta1\lib\net45\Autofac.dll</HintPath>
     </Reference>
-    <Reference Include="FluentAssertions, Version=5.4.2.0, Culture=neutral, PublicKeyToken=33f2691a05b67b6a, processorArchitecture=MSIL">
-      <HintPath>..\packages\FluentAssertions.5.4.2\lib\net45\FluentAssertions.dll</HintPath>
+    <Reference Include="Castle.Core, Version=4.0.0.0, Culture=neutral, PublicKeyToken=407dd0808d44fbdc, processorArchitecture=MSIL">
+      <HintPath>..\packages\Castle.Core.4.3.1\lib\net45\Castle.Core.dll</HintPath>
+    </Reference>
+    <Reference Include="FluentAssertions, Version=5.5.3.0, Culture=neutral, PublicKeyToken=33f2691a05b67b6a, processorArchitecture=MSIL">
+      <HintPath>..\packages\FluentAssertions.5.5.3\lib\net45\FluentAssertions.dll</HintPath>
+    </Reference>
+    <Reference Include="NSubstitute, Version=3.1.0.0, Culture=neutral, PublicKeyToken=92dd2e9066daa5ca, processorArchitecture=MSIL">
+      <HintPath>..\packages\NSubstitute.3.1.0\lib\net46\NSubstitute.dll</HintPath>
     </Reference>
     <Reference Include="nunit.framework, Version=3.11.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
       <HintPath>..\packages\NUnit.3.11.0\lib\net45\nunit.framework.dll</HintPath>
     </Reference>
     <Reference Include="System" />
+    <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
     <Reference Include="System.Drawing" />
+    <Reference Include="System.Threading.Tasks.Extensions, Version=4.1.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
+      <HintPath>..\packages\System.Threading.Tasks.Extensions.4.3.0\lib\portable-net45+win8+wp8+wpa81\System.Threading.Tasks.Extensions.dll</HintPath>
+    </Reference>
     <Reference Include="System.ValueTuple, Version=4.0.2.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
       <HintPath>..\packages\System.ValueTuple.4.4.0\lib\net461\System.ValueTuple.dll</HintPath>
     </Reference>
+    <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -58,17 +69,23 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CircularCloudLayouter_Should.cs" />
-    <Compile Include="MathHelper.cs" />
+    <Compile Include="Layouter\CircularCloudLayouter_Should.cs" />
     <Compile Include="MathHelper_Should.cs" />
-    <Compile Include="PointExtension_Should.cs" />
+    <Compile Include="Layouter\PointExtension_Should.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RectangleExtension_Should.cs" />
-    <Compile Include="Spiral_Should.cs" />
-    <Compile Include="WordsCloudLayouter_Should.cs" />
-    <Compile Include="WordsCloudVisualizer_Should.cs" />
-    <Compile Include="Word_Should.cs" />
+    <Compile Include="Layouter\RectangleExtension_Should.cs" />
+    <Compile Include="Layouter\Spiral_Should.cs" />
+    <Compile Include="Layouter\WordsCloudBuilder_Should.cs" />
+    <Compile Include="Visualizer\WordsCloudVisualizer_Should.cs" />
+    <Compile Include="WordProcessing\BoringWordsFilter_Should.cs" />
+    <Compile Include="WordProcessing\FrequencyWeighter_Should.cs" />
+    <Compile Include="WordProcessing\TextPreprocessor_Should.cs" />
+    <Compile Include="WordProcessing\TxtReader_Should.cs" />
+    <Compile Include="WordProcessing\WeightedWord_Should.cs" />
+    <Compile Include="WordProcessing\WordsChanger_Should.cs" />
+    <Compile Include="WordProcessing\WordSizeConverter_Should.cs" />
+    <Compile Include="Layouter\Word_Should.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
@@ -79,6 +96,11 @@
       <Project>{baa65ed5-cd11-4dd9-98ac-49ff3b6b28e9}</Project>
       <Name>TagsCloudVisualization</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="WordProcessing\testFile.txt">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
