--- TagsCloudVisualization_Tests/WordsCloudLayouter_Should.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using FluentAssertions;
-using NUnit.Framework;
-using TagsCloudVisualization;
-
-namespace TagsCloudVisualization_Tests
-{
-    [TestFixture]
-    public class WordsCloudLayouter_Should
-    {
-        [Test]
-        [Ignore("Not Implemented")]
-        public void LayWords_AndReturnWordEntity()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
