--- TagsCloudVisualization_Tests/WordsCloudVisualizer_Should.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using FluentAssertions;
-using NUnit.Framework;
-using TagsCloudVisualization;
-
-namespace TagsCloudVisualization_Tests
-{
-    [TestFixture]
-    public class WordsCloudVisualizer_Should
-    {
-        [Test]
-        [Ignore("Not Implemented")]
-        public void DrawRectangles_BeCorrectSize()
-        {
-           throw new NotImplementedException();
-        }
-    }
-}
