+++ TagsCloudVisualization_Tests/packages.config
@@ -1,7 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="FluentAssertions" version="5.4.2" targetFramework="net461" />
+  <package id="Castle.Core" version="4.3.1" targetFramework="net461" />
+  <package id="FluentAssertions" version="5.5.3" targetFramework="net461" />
+  <package id="NSubstitute" version="3.1.0" targetFramework="net461" />
   <package id="NUnit" version="3.11.0" targetFramework="net461" />
   <package id="System.Collections.Immutable" version="1.3.0" targetFramework="net461" />
+  <package id="System.Threading.Tasks.Extensions" version="4.3.0" targetFramework="net461" />
   <package id="System.ValueTuple" version="4.4.0" targetFramework="net461" />
 </packages>
